Task 5 Controller Class Converter , Exceptions added
Store is still in progress.

+++ src/main/java/com/sigma/sigmastore/service/GoodsService.java
@@ -1,7 +1,10 @@
 package com.sigma.sigmastore.service;
 
+import com.sigma.sigmastore.converter.GoodsConverter;
 import com.sigma.sigmastore.dao.GoodsDao;
 import com.sigma.sigmastore.domain.Goods;
+import com.sigma.sigmastore.exceptions.DataNotFoundException;
+import com.sigma.sigmastore.exceptions.SymanticException;
 import com.sigma.sigmastore.repository.GoodsRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,17 +14,53 @@ import org.springframework.stereotype.Service;
 public class GoodsService {
 
 	private GoodsRepository goodsRepository;
+	private GoodsConverter goodsConverter;
+
+	public GoodsDao getGoods(int id) {
+		Goods returnedgood = goodsRepository.getGoods(id);
+		return new GoodsDao(returnedgood.getName(), returnedgood.getPrice(), returnedgood.getUnits());
+	}
 
 	public GoodsDao createGoods(GoodsDao goodsDao) {
-		Goods createGoods = goodsRepository.createGoods(new Goods(goodsDao.getName(),
-				goodsDao.getPrice(), goodsDao.getUnits()));
+		validate(goodsDao.getPrice() < 1, "Price Must be Positive");
+		validate(goodsDao.getUnits() < 0, "Units Must be more than zero");
 
-		return new GoodsDao(createGoods.getName(), createGoods.getPrice(), createGoods.getUnits());
+		Goods createGoods = goodsRepository.createGoods(goodsConverter.fromDao(goodsDao));
+
+		return goodsConverter.fromDomain(createGoods);
 	}
 
-	public void deleteGoods(Integer id) {
-		goodsRepository.deleteGoods(id);
+	public GoodsDao updateGoods(int id, GoodsDao goodsDao) {
+		CheckExisting(id);
+		validate(goodsDao.getPrice() < 1, "Price Must be Positive");
+		validate(goodsDao.getUnits() < 0, "Units Must be more than zero");
 
+		Goods updateGoods = goodsRepository.updateGoods(id,goodsConverter.fromDao(goodsDao));
+
+		return goodsConverter.fromDomain(updateGoods);
 	}
 
+	public void deleteGoods(int id) {
+		Goods goods = goodsRepository.getGoods(id);
+		if (goods != null) {
+			goodsRepository.deleteGoods(id);
+		}else{
+			throw new DataNotFoundException("Book with id = " + id + " is not found" );
+		}
+	}
+
+	private void CheckExisting(int id) {
+		Goods goods = goodsRepository.getGoods(id);
+		if (goods != null) {
+			System.out.println("The goods With id " + id + "is exist");
+		}else{
+			throw new DataNotFoundException("Book with id = " + id + " is not found" );
+		}
+	}
+
+	private void validate(boolean goodsDao, String message) {
+		if (goodsDao){
+			throw new SymanticException(message);
+		}
+	}
 }
